% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR95CI_function.R
\name{OR_95CI}
\alias{OR_95CI}
\title{OR and 95% confidence interval}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{a vector of coefficients from a logistic regression model}

\item{se}{a vector of standard errors from a logistic regression model}

\item{siglevel}{the significance level e.g. .05}

\item{roundto}{number of digits to round to}
}
\value{
ORresult
}
\description{
a function that returns the odds ratios and 95% confidence intervals from a vector of coefficient estimates and a vector of standard errors
}
\examples{
OR_95CI(0.5, 0.1, 0.05, 2)
}
\author{
Meagan Lacroix
}
