% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toydata}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{y}{binary outcome}
\item{x1}{continuous predictor}
\item{x2}{continuous predictor}
}
}
\usage{
toydata
}
\description{
Example data to test R function
}
\keyword{datasets}
